# Optimize NamespacedKeyword (and NamespacedSymbol)