--- a/edn/Cargo.toml
+++ b/edn/Cargo.toml
@@ -20,6 +20,10 @@ uuid = { version = "0.5", features = ["v4", "serde"] }
 serde = { version = "1.0", optional = true }
 serde_derive = { version = "1.0", optional = true }
 
+[dev-dependencies]
+serde_test = "1.0"
+serde_json = "1.0"
+
 [features]
 serde_support = ["serde", "serde_derive"]