--- a/src/lib.rs
+++ b/src/lib.rs
@@ -76,18 +76,18 @@ macro_rules! var {
 macro_rules! kw {
     ( : $ns:ident / $n:ident ) => {
         // We don't need to go through `new` -- `ident` is strict enough.
-        $crate::NamespacedKeyword {
-            namespace: stringify!($ns).into(),
-            name: stringify!($n).into(),
-        }
+        $crate::NamespacedKeyword::new(
+            stringify!($ns),
+            stringify!($n)
+        )
     };
 
     ( : $ns:ident$(. $nss:ident)+ / $n:ident ) => {
         // We don't need to go through `new` -- `ident` is strict enough.
-        $crate::NamespacedKeyword {
-            namespace: concat!(stringify!($ns) $(, ".", stringify!($nss))+).into(),
-            name: stringify!($n).into(),
-        }
+        $crate::NamespacedKeyword::new(
+            concat!(stringify!($ns) $(, ".", stringify!($nss))+),
+            stringify!($n)
+        )
     };
 }