@rnewman 2018-05-10
I’m wrapping this up and taking it further to support aliasing in pull.