@ncalexan 2018-05-10
> I’m wrapping this up and taking it further to support aliasing in pull.

I'm glad to see this landing.  Could this help when consuming `StructuredMap`?  Right now the keys are `ValueRc<NamespacedKeyword>`, which is frustrating to create.